Prepare for distribution. (2.1.0b)
The snowflake module has been move to its own package with accompanying
tests. A manifest and a description has been added to the top-level.
The build script has been updated to a format used in the Python PUG
tutorial and uses setuptools instead of distutils now. Also, the README
will be written in reStructredText and now has a file extension.