Continue preparing for distribution. (3.0.0b)
A few major changes/upgrades have been introduced including a test
module and a new interface to accommodate API key handling. It tests
each method extensively using dependency injection and example response
data while supporting the new interface which has been modified because
OGRe is now a class. API keys are attached to retriever objects in the
constructor so as to avoid having a separate credentials module which
was inconvenient, unportable, and inelegant. Documentation has seen
some minor improvements but will be created with Sphinx soon (before
distribution on PyPI).