Add runtime modifiers. (4.3.0b)
Many parameters are now accepted to affect runtime functionality.
`test` specifies a trial run and `test_message`, if specified, will log
a description about the trial. Since Tweet text is already on hand (as
opposed to images which are retrieved manually after Twitter’s
response), it is always returned (even if “text” is not specified in
`media`) unless `strict_media` is set to True. Since all queries on the
Twitter Search API must be HTTPS, images are now also fetched with
HTTPS by default unless `secure` is set to False. The test suite for
Twitter has been overhauled and now does not depend on an internet
connection. The `network` kwargs can be passed to analyze how images
are retrieved.